refactor: group ethereum and solana modules into src/chains
