+++ frontend/src/chains/solana/useDynamicSolanaWallet.ts
