--- frontend/src/contexts/ethereum/index.ts
@@ -1,1 +1,0 @@
-// Ethereum-specific contexts (empty for now, kept for structure)
