--- frontend/src/contexts/index.ts
@@ -1,3 +1,0 @@
-// Re-export all contexts from organized structure
-export * from './ethereum';
-export * from './solana';
